Added a "CannotParseException", which is thrown if a property value cannot be parsed.  This will cause the property to be unmarshalled as a "RawType" instead (or as an "XmlType" with xCards).
A "CannotParseException" is now consistently thrown during the parsing of xCard properties if the parser cannot find the XML element which holds the property's value.  Before, in this situation, a "SkipMeException" was sometimes thrown, but not always.

If a TZ property in an xCard contains both a <text> and a <uri> element, only the <text> element will be parsed.  Before, both were parsed, but the standard does not allow for both to be defined together in the same property.