Made the parsing of folded, quoted-printable lines more robust.
The FoldingLineReader class was merged into the VCardRawReader class.
Instead of reading the data one line at a time, and then parsing the
vCard components out of that line, it now read the data one character at
a time, unfolding lines and parsing the vCard syntax as it goes.

This was done in order to make the unfolding of folded, quoted-printable
lines more robust.  To determine if a line was quoted-printable, it used
to simply look for the string "quoted-printable" somewhere before the
property value.  However, this technique would fail if (1) that exact
string happened to be part of an arbitrary parameter value or (2) if the
parameters themselves were folded, and the "quoted-printable" parameter
was not on the first folded line.  The changes in this commit fix these
two issues.