Changes regarding how invalid characters are handled in VCardRawWriter.
If a parameter name or value contains any characters which can't be
escaped or encoded, and which break the vCard syntax, a runtime
exception is now thrown.  Before, parameter names were not checked at
all.  And invalid characters in parameter values were either silently
removed or silently replaced with a similar character.

Spec-compliance for correct character usage in parameter names and
values is now checked in "VCardParameters.validate()".