Made the ADR property parser multi-value aware.
According to the ADR property syntax, it is possible for each component
of the address to contain multiple values.  ez-vcard would previously
only recognize the first value and ignore the rest.  This commit changes
this so that ez-vcard holds onto all the values.

This becomes an issue when the vCard contains a comma character in the
street address component, and that comma is not properly escaped. If
this happened, everything after the comma would be lost because ez-vcard
would ignore everything after the comma.  This commit resolves this
issue by storing all values in a component instead of just the first
value.

To help with unescaped commas in street addresses, a
"getStreetAddressFull" method was added to the "Address" property class,
which joins all the street address values with a comma delimitor.