"getTypes()" methods now return a List that is backed by the parameters.
Since we're using a List instead of a Set, order is now preserved when
multiple TYPE values are specified. However, duplicates can now be
entered, so care must be taken by the developer not to insert
duplicates.

Removed the "addType" and "removeType" methods because values can be
added and removed to/from the List object.