Commas should not be escaped in version 2.1.
Version 2.1 does not recognize multi-valued components the way 3.0 and
4.0 do. For example, with the N property, it is not possible to define
multiple prefixes in version 2.1. Therefore, the 2.1 specification does
not specify any need to escape comma characters and thusly, they should
not be escaped.

Validation warnings were added to the N and ADR properties, stating that
multi-valued fields are not recognized in version 2.1.

https://github.com/mangstadt/ez-vcard/issues/67