Allow dates with single digit months/days to be parsed
A "quite common" CardDAV server formats BDAY properties with single
digit months and days. This caused ez-vcard to parse the property as a
RawProperty because this is not spec-compliant.

This commit allows the month and/or date to be single digit, provided
the date uses extended format (removing the dashes can result in
ambiguity for certain dates such as Nov 3).

https://github.com/mangstadt/ez-vcard/issues/102